Render the scale domain as a path.
The appearance of the domain path can be customized using the "mode" property.
If "open", then the domain path is a straight line; if "closed", then the domain
path has ticks at the start and end to indicate the extent. In closed mode, if
a tick is one of the domain extents, then the tick's line is hidden to avoid
overlap with the path domain.